Move api to external folder since it's not our api
